refactor: sortObjectKeys
