+++ packages/shared/src/utils/file-utils.spec.ts
@@ -1,4 +1,4 @@
-import { setNestedProperty, sortJsonKeys } from "./file-utils";
+import { setNestedProperty } from "./file-utils";
 
 describe("setNestedProperty", () => {
   it("Sets object deep property", () => {
@@ -14,23 +14,5 @@ describe("setNestedProperty", () => {
     const varProperty = "c";
     const res = setNestedProperty<any>(`a.b.${varProperty}`, 1);
     expect(res).toEqual({ a: { b: { c: 1 } } });
-  });
-});
-
-describe("sortJsonKeys", () => {
-  it("Sorts nested json by key", () => {
-    const input = {
-      b: "foo",
-      c: null,
-      a: {
-        f: 6,
-        e: 5,
-      },
-      d: [],
-    };
-    const res = sortJsonKeys(input);
-    expect(Object.keys(res)).toEqual(["a", "b", "c", "d"]);
-    expect(Object.keys(res.a)).toEqual(["e", "f"]);
-    expect(Object.values(res.a)).toEqual([5, 6]);
   });
 });
