+++ packages/shared/src/utils/index.ts
@@ -3,5 +3,6 @@ export * from "./cli-utils";
 export * from "./delimiters";
 export * from "./file-utils";
 export * from "./logging.utils";
+export * from "./object-utils";
 export * from "./string-utils";
 export * from "./typescript-utils";
